добавляет модуль формы
